requirements.txt
folium==0.20.0
haversine==2.9.0
numpy==2.4.2
pandas==3.0.0
Pillow==12.1.0
plotly==5.24.1
streamlit==1.44.1
streamlit_folium==0.26.1
